Extract Yesod.Worker module
in preparation for packaging it separately. The API as used by
Foundation / Application is solid, but the Jobs / Jobs.Types are stil
too tightly coupled. I'd like to figure out a good way for the Job type
to be a type associated with the YesodWorker instance, and to move that
instance into Foundation.hs.